Moved components to other files, removed comments
